Updated API models
