--- prime-angular-frontend/src/app/features/enrolment/pages/contact/contact.model.ts
@@ -1,8 +1,0 @@
-export interface Contact {
-  prefersEmail: boolean;
-  contactEmail: string;
-  prefersPhone: boolean;
-  contactPhone: string;
-  voicePhone: string;
-  voiceExtension: number;
-}
