--- prime-angular-frontend/src/app/features/enrolment/pages/pharmanet-access/pharmanet-access.model.ts
@@ -1,12 +1,0 @@
-export interface PharmanetAccess {
-  organizations: Organizations[];
-}
-
-interface Organizations {
-  id: number;
-  name: string;
-  organizationTypeCode: string;
-  city: string;
-  startDate: Date;
-  endDate: string;
-}
