+++ prime-angular-frontend/src/app/features/enrolment/pages/professional-info/professional-info.component.html
@@ -80,7 +80,7 @@
     </mat-radio-button>
   </mat-radio-group>
 
-  <p>Are you accessing PharmNet on behalf of someone else?</p>
+  <p>Are you accessing PharmaNet on behalf of someone else?</p>
 
   <mat-radio-group aria-labelledby="example-radio-group-label"
                    class="example-radio-group"
