--- prime-angular-frontend/src/app/features/enrolment/pages/profile/profile.model.ts
@@ -1,13 +1,0 @@
-import { Address } from '../../shared/models/address.model';
-
-export interface Profile {
-  firstName: string;
-  middleName?: string;
-  lastName: string;
-  dateOfBirth: Date;
-  preferredFirstName?: string;
-  preferredMiddleName?: string;
-  preferredLastName?: string;
-  physicalAddress: Address;
-  mailingAddress?: Address;
-}
