+++ prime-angular-frontend/src/app/features/enrolment/shared/models/college-certification.model.ts
@@ -1,0 +1,8 @@
+export interface CollegeCertification {
+  id: number;
+  collegeCode: number;
+  licenseNumber: string;
+  licenseCode: number;
+  renewalDate: Date;
+  practiceCode?: number;
+}
