+++ prime-angular-frontend/src/app/features/enrolment/shared/models/enrolment.model.ts
@@ -1,13 +1,41 @@
-import { Profile } from '../../pages/profile/profile.model';
-import { Contact } from '../../pages/contact/contact.model';
-import { ProfessionalInfo } from '../../pages/professional-info/professional-info.model';
-import { SelfDeclaration } from '../../pages/self-declaration/self-declaratin.model';
-import { PharmanetAccess } from '../../pages/pharmanet-access/pharmanet-access.model';
+import { Moment } from 'moment';
+import { Address } from './address.model';
+import { CollegeCertification } from './college-certification.model';
+import { Job } from './job.model';
+import { Organization } from './organization.model';
 
 export interface Enrolment {
-  profile: Profile;
-  contact: Contact;
-  professionalInfo: ProfessionalInfo;
-  selfDeclaration: SelfDeclaration;
-  pharmanetAccess: PharmanetAccess;
+  enrollee: {
+    userId: string;
+    firstName: string;
+    middleName: string;
+    lastName: string;
+    preferredFirstName: string;
+    preferredMiddleName: string;
+    preferredLastName: string;
+    dateOfBirth: Moment;
+    physicalAddress: Address,
+    mailingAddress: Address,
+    contactEmail: string;
+    contactPhone: string;
+    voicePhone: string;
+    voiceExtension: string;
+  };
+  hasCertification: boolean;
+  certifications: CollegeCertification[];
+  isDeviceProvider: boolean;
+  deviceProviderNumber: string;
+  isInsulinPumpProvider: boolean;
+  isAccessingPharmaNetOnBehalfOf: boolean;
+  jobs: Job[];
+  hasConviction: boolean;
+  hasRegistrationSuspended: boolean;
+  hasDisciplinaryAction: boolean;
+  hasPharmaNetSuspended: boolean;
+  organizations: Organization[];
+}
+
+export interface SelfDeclarationIncident {
+  details: string;
+  documents: any[];
 }
