+++ prime-angular-frontend/src/app/features/enrolment/shared/models/job.model.ts
@@ -1,0 +1,4 @@
+export interface Job {
+  id?: number;
+  title: string;
+}
