+++ prime-angular-frontend/src/app/features/enrolment/shared/models/organization.model.ts
@@ -1,0 +1,10 @@
+import { Moment } from 'moment';
+
+export interface Organization {
+  id?: number;
+  name: string;
+  organizationTypeCode: string;
+  city: string;
+  startDate: Moment;
+  endDate: Moment;
+}
