+++ prime-angular-frontend/src/scss/_font.scss
@@ -2,7 +2,6 @@
 // Theme Fonts
 //
 
-// TODO: Update portal to use the same font as the site
 @import url('https://fonts.googleapis.com/css?family=Open+Sans:300,400|Roboto');
 
 //
